requirements.txt
docker
apache-airflow
pandas
sqlalchemy
google-api-client
json
requests
